Fix index tag mismatch.
This is an incredibly convoluted fix, based on putting the array in variable parameters.  I also have to use a dummy array variable to copy the address over to get "SREF" correct and because the old compiler crashes when I try just using assembly to call "_SortDeepArray".  I've also added "size" as a variable parameter, because i find I normally only want to sort a subset of the array according to how many elements I have actually written to: ".size = upperbound".